Rework task organization & references
