+++ package.json
@@ -86,6 +86,7 @@
     "stream-combiner": "0.2.2",
     "stylelint": "4.5.1",
     "through2": "2.0.1",
+    "uri.js": "^0.1.3",
     "vinyl-fs": "2.4.2",
     "webdriverjs-angular": "0.1.3",
     "yargs": "4.2.0"
