+++ tools/misc/run-command.js
