+++ tools/misc/stream-cache.js
